refactor(buider-component): refactor builder-component
"refac: #48"
