--- template/component/Component.accessibility.js
@@ -1,11 +1,0 @@
-import React from 'react'
-import { render } from '@testing-library/react'
-import { axe, toHaveNoViolations } from 'jest-axe'
-
-import { Default } from './Componet.stories'
-expect.extend(toHaveNoViolations)
-
-it('renders avatar default', async () => {
-  const { container } = render(<Default {...Default.args} />)
-  expect(await axe(container)).toHaveNoViolations()
-})
