+++ template/component/Component.module.css
@@ -1,3 +1,3 @@
-.container {
+.component {
   display: flex;
 }
