+++ template/component/index.js
@@ -1,3 +1,3 @@
-export { default, Componet } from './Componet'
+export { default, Component } from './Component'
 export { options } from './constants'
-export { default as styles } from './Componet.module.css'
+export { default as styles } from './Component.module.css'
